review of wif creation with suggestions and comments
