+++ cmd/ocm/gcp/delete-wif-config.go
@@ -113,3 +113,11 @@ func deleteWorkloadIdentityPool(ctx context.Context, gcpClient gcp.GcpClient, wi
 	log.Printf("Workload identity pool %s deleted", poolName)
 	return nil
 }
+
+func generateServiceAccountID(serviceAccount models.ServiceAccount) string {
+	serviceAccountID := "z-" + serviceAccount.Id
+	if len(serviceAccountID) > 30 {
+		serviceAccountID = serviceAccountID[:30]
+	}
+	return serviceAccountID
+}
