+++ pkg/alpha_ocm/client.go
@@ -1,0 +1,5 @@
+package alphaocm
+
+type OcmClient interface {
+	CreateWifConfig(wifConfigInput interface{}) error
+}
