+++ pkg/alpha_ocm/connection.go
@@ -1,0 +1,3 @@
+package alphaocm
+
+// TODO Create a connection to an ocm-sdk fork containing wif API
